pequeños cambios y prueba de funcionalidades

+++ src/IAccionDinero/IAccionDinero.java
@@ -2,5 +2,5 @@ package IAccionDinero;
 import Jugador.Jugador;
 
 public interface IAccionDinero {
-    public double accionDinero(Jugador jugador);
+    public int accionDinero(Jugador jugador);
 }
